+++ src/vs/editor/contrib/hover/test/browser/contentHover.test.ts
@@ -7,7 +7,7 @@ import assert from 'assert';
 import { ensureNoDisposablesAreLeakedInTestSuite } from 'vs/base/test/common/utils';
 import { Position } from 'vs/editor/common/core/position';
 import { Range } from 'vs/editor/common/core/range';
-import { RenderedHover } from 'vs/editor/contrib/hover/browser/contentHoverController';
+import { RenderedContentHover } from 'vs/editor/contrib/hover/browser/contentHoverRendered';
 import { IHoverPart } from 'vs/editor/contrib/hover/browser/hoverTypes';
 import { TestCodeEditorInstantiationOptions, withTestCodeEditor } from 'vs/editor/test/browser/testCodeEditor';
 
@@ -18,7 +18,7 @@ suite('Content Hover', () => {
 	test('issue #151235: Gitlens hover shows up in the wrong place', () => {
 		const text = 'just some text';
 		withTestCodeEditor(text, {}, (editor) => {
-			const actual = RenderedHover.computeHoverRanges(
+			const actual = RenderedContentHover.computeHoverPositions(
 				editor,
 				new Range(5, 5, 5, 5),
 				[<IHoverPart>{ range: new Range(4, 1, 5, 6) }]
@@ -38,7 +38,7 @@ suite('Content Hover', () => {
 		const text = 'just some text';
 		const opts: TestCodeEditorInstantiationOptions = { wordWrap: 'wordWrapColumn', wordWrapColumn: 6 };
 		withTestCodeEditor(text, opts, (editor) => {
-			const actual = RenderedHover.computeHoverRanges(
+			const actual = RenderedContentHover.computeHoverPositions(
 				editor,
 				new Range(1, 8, 1, 8),
 				[<IHoverPart>{ range: new Range(1, 1, 1, 15) }]
